requirements.txt
numpy
PyP100
meteostat
python-dotenv
